Fix boundary issue in findrange
