+++ banyand/stream/block_test.go
@@ -427,99 +427,6 @@ func Test_marshalAndUnmarshalBlock(t *testing.T) {
 	}
 }
 
-func Test_findRange(t *testing.T) {
-	type args struct {
-		timestamps []int64
-		min        int64
-		max        int64
-		includeMin bool
-		includeMax bool
-	}
-	tests := []struct {
-		name      string
-		args      args
-		wantStart int
-		wantEnd   int
-		wantExist bool
-	}{
-		{
-			name: "Test with empty timestamps",
-			args: args{
-				timestamps: []int64{},
-				min:        1,
-				max:        10,
-			},
-			wantStart: -1,
-			wantEnd:   -1,
-			wantExist: false,
-		},
-		{
-			name: "Test with single timestamp",
-			args: args{
-				timestamps: []int64{1},
-				min:        1,
-				max:        1,
-				includeMin: true,
-				includeMax: true,
-			},
-			wantStart: 0,
-			wantEnd:   0,
-			wantExist: true,
-		},
-		{
-			name: "Test with range not in timestamps",
-			args: args{
-				timestamps: []int64{1, 2, 3, 4, 5},
-				min:        6,
-				max:        10,
-			},
-			wantStart: -1,
-			wantEnd:   -1,
-			wantExist: false,
-		},
-		{
-			name: "Test with range in timestamps",
-			args: args{
-				timestamps: []int64{1, 2, 3, 4, 5},
-				min:        2,
-				max:        4,
-				includeMin: true,
-				includeMax: true,
-			},
-			wantStart: 1,
-			wantEnd:   3,
-			wantExist: true,
-		},
-		{
-			name: "Test with range as timestamps",
-			args: args{
-				timestamps: []int64{1, 2, 3, 4, 5},
-				min:        1,
-				max:        5,
-				includeMin: true,
-				includeMax: true,
-			},
-			wantStart: 0,
-			wantEnd:   4,
-			wantExist: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			gotStart, gotEnd, gotExist := findRange(tt.args.timestamps, tt.args.min, tt.args.max, tt.args.includeMin, tt.args.includeMax)
-			if gotStart != tt.wantStart {
-				t.Errorf("findRange() gotStart = %v, want %v", gotStart, tt.wantStart)
-			}
-			if gotEnd != tt.wantEnd {
-				t.Errorf("findRange() gotEnd = %v, want %v", gotEnd, tt.wantEnd)
-			}
-			if gotExist != tt.wantExist {
-				t.Errorf("findRange() gotExist = %v, want %v", gotExist, tt.wantExist)
-			}
-		})
-	}
-}
-
 func Test_blockPointer_append(t *testing.T) {
 	type fields struct {
 		timestamps  []int64
